+++ docs/changelog.md
@@ -38,6 +38,7 @@ MML changes:
  * Added `_M` channel transpose and `__M` adjust channel transpose MML commands
  * Added `#OldTranspose` header to restore the v0.1.1 and earlier transpose behaviour
    by silently converting `_` and `__` commands to `_M` and `__M`.
+ * Added `_{}` change key signature command
 
 
 Version 0.1.1